// src/Twin/Scenes/Enums/GroupEnums.ts
export enum GroupTypes {
  PLAYER = "player",
  GHOST_PLAYER = "ghost player",
  GROUND = "ground",
  PLAYER_GROUND = "player ground",
  GHOST_GROUND = "ghost ground",
  ENEMY = "enemy",
  PLAYER_ENEMY = "player enemy",
  GHOST_ENEMY = "ghost enemy",
  COIN = "coin",
  PLAYER_COIN = "player coin",
  GHOST_COIN = "ghost coin",
}